GH Actions: start testing against PHP 8.1
While PHP 8.1 has not been released yet, it is common to allow such a build to fail.

The tests currently fail due to PHP 8.1 incompatibilities in Mockery.
A PR for (most of) these is already open in the Mockery repo: https://github.com/mockery/mockery/pull/1130

Once Mockery has been made compatible with PHP 8.1, the `continue-on-error` setting should be revisited after verifying that all tests pass.