Tests: introduce a base TestCase class
This base test class only contains two methods:
* A custom assertion to allow comparing two text strings with different line endings and consider those the same.
* A helper method to normalize line endings in arbitrary text strings.

This is a first step towards fixing the test failures in the Windows builds.

Most of the test failures in the Windows builds are due to the git check out of the code being done with `core.autocrlf true` and the line endings in the "expected" value not matching the line endings in the "actual" value.

For more detailed information, see issue 287